Move getActiveAccountNumbersForIdentityId to syncUser.ts
Relocated the getActiveAccountNumbersForIdentityId function and its related schema from getAccountsForIdentityId.ts to syncUser.ts for better module organization. Removed the now-unused code from getAccountsForIdentityId.ts.
