+++ modules/sync-supporter-product-data/src/syncUser.ts
@@ -4,11 +4,11 @@ import { getProductCatalogFromApi } from '@modules/product-catalog/api';
 import { ProductCatalogHelper } from '@modules/product-catalog/productCatalog';
 import type { Stage } from '@modules/stage';
 import { zuoraDateFormat } from '@modules/zuora/common';
-import { getActiveAccountNumbersForIdentityId } from '@modules/zuora/getAccountsForIdentityId';
 import { getSubscriptionsByAccountNumber } from '@modules/zuora/getSubscription';
 import { ZuoraClient } from '@modules/zuora/zuoraClient';
 import type { RatePlan, ZuoraSubscription } from '@modules/zuora/zuoraSchemas';
 import dayjs from 'dayjs';
+import { z } from 'zod';
 
 type MessageBody = {
 	subscriptionName: string;
@@ -107,4 +107,28 @@ const sendToQueue = async (stage: Stage, message: MessageBody) => {
 	const response = await client.send(command);
 	console.log(`Response from message send was ${JSON.stringify(response)}`);
 	return response;
+};
+
+const responseSchema = z.object({
+	records: z.array(z.object({ AccountNumber: z.string() })),
+});
+
+type Response = z.infer<typeof responseSchema>;
+
+export const getActiveAccountNumbersForIdentityId = async (
+	zuoraClient: ZuoraClient,
+	identityId: string,
+) => {
+	const path = `/v1/action/query`;
+
+	const body = {
+		queryString: `select accountNumber from account where status = 'Active' and IdentityId__c = '${identityId}'`,
+	};
+	const response: Response = await zuoraClient.post(
+		path,
+		JSON.stringify(body),
+		responseSchema,
+	);
+
+	return response.records.map((record) => record.AccountNumber);
 };
